some minor argument issues fixed

+++ src/HaloMerged.cpp
@@ -79,14 +79,14 @@ bool HaloMerged::Halo_Init(matrix<double>& A)
 		int n_recv = recvLength[i];
 		//made up some syntax to deal with edges. getEdge is implemented for conceptual purposes
 		//and will not actually work
-		MPI_Irecv(A.getEdge(i), n_recv, MPI_LONG, neighbors[i], tag, MPI_COMM_WORLD, requests+i); 
+		MPI_Irecv(A.getEdge(i), n_recv, MPI_DOUBLE, neighbors[i], tag, MPI_COMM_WORLD, requests+i); 
 	}
 
 	//Sends
 	for(int i=0; i<numNeighbors; i++)
 	{
 		int n_send = sendLength[i];
-		MPI_Send(A.getExternal(i), n_send, MPI_LONG, neighbors[i], tag, MPI_COMM_WORLD);
+		MPI_Send(A.getExternal(i), n_send, MPI_DOUBLE, neighbors[i], tag, MPI_COMM_WORLD);
 	}
 	MPI_Waitall(numNeighbors, requests, status);
